@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { User } from '../user';
 import { UserService } from '../user.service';
+import { User } from '../user';
 
 @Component({
   selector: 'app-sign-up',
